Rearrange the file
